major refactor of uxds_acct_add()
Changed how the LDAPMod arrays of structs are filled in. Instead of
using char*[]'s and swapping values back and forth, they are now
filled in by an array of uxds_attr_t structs, which is more efficient.
This is the first phase of a major refactor that will extend to
uxds_acct_mod() and uxds_sudo_(add|mod)() functions next.