reorganized code to fixed broken password change/expire operations
It was formerly not possible to perform a krb5 password change or
expiry and modify user attributes simultaneously. This was fixed
by eliminating the skipmod goto and reorganizing the code in a more
sane fashion.  The password changes happen first, and if there are
no modifications to be done, uxds_acct_mod returns.  If there are,
the code continues normally to perform the mods.